style(ml): ruff format + lint conformance for Phase 1 research probes and Layer-2 adviser
Non-behavioral conformance pass on the modules this branch introduced, so
the CI ruff-format/ruff-check gate stays green:

src/nexus_scalp/research/{aggregate,champion_ceiling,contract_check,dataset_stats}.py
  (from 09077764 research(phase0)): ruff format; fix I001 import blocks,
  F401/F841 unused imports+locals, RUF046 redundant int() casts, B905
  zip() without strict=.

src/nexus_scalp/position_adviser/ (from 8d00483f TASK-POSA): format +
  the same lint classes; __init__.py F401 is a documented public
  re-export (PositionAdvisory is in __all__ and consumed via the package)
  so it takes a noqa, not a deletion.

Verified: ruff format --check and ruff check clean on all touched files;
Phase-1 + position-adviser suites 100 passed / 1 skipped (unchanged
verdicts). mypy is CI-continue-on-error; the 16 remaining errors are
pre-existing at HEAD (27 at HEAD before this pass) and live in regions
this pass did not touch.

+++ src/nexus_scalp/position_adviser/integration.py
@@ -139,7 +139,7 @@ def apply_advisory_to_hold_score(
 
     max_pen = float(getattr(getattr(service, "config", None), "max_hold_score_penalty", 25.0))
     clamped = max(-max_pen, min(0.0, adj))
-    new_score = max(0, min(100, int(round(hold_score + clamped))))
+    new_score = max(0, min(100, round(hold_score + clamped)))
 
     if new_score == hold_score:
         return hold_score, advisory.to_dict()
